services/api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.37
asyncpg==0.30.0
psycopg2-binary==2.9.10
redis==5.2.1
httpx==0.28.1
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
email-validator==2.2.0
numpy==1.26.4
pandas==2.2.3
scipy==1.13.1
scikit-learn==1.5.2
scikit-surprise==1.1.4
xgboost==2.1.3
